put the tasklist component in its own file
